+++ src/util/browserErrorPlugin.js
@@ -1,22 +1,21 @@
-/* eslint no-var: 0 */
-var path = require('path');
+const path = require('path');
 
-var BrowserErrorPlugin = function() {};
+const BrowserErrorPlugin = function() {};
 BrowserErrorPlugin.prototype.apply = function(compiler) {
     compiler.plugin('done', function(stats) {
         if (stats.hasErrors()) {
-            var nonLinterErrors = stats.compilation.errors.filter(function(err) {
+            const nonLinterErrors = stats.compilation.errors.filter(function(err) {
                 return err &&
                     err.module &&
                     err.module.issuer &&
                     err.module.issuer.indexOf('eslint-loader') === -1;
             });
             if (nonLinterErrors.length) {
-                var outputFileSystem = compiler.outputFileSystem;
-                var outputOptions = compiler.options.output;
-                var main = path.join(outputOptions.path, outputOptions.filename);
-                var errors = nonLinterErrors.map(function(err) {
-                    var msg = '';
+                const outputFileSystem = compiler.outputFileSystem;
+                const outputOptions = compiler.options.output;
+                const main = path.join(outputOptions.path, outputOptions.filename);
+                const errors = nonLinterErrors.map(function(err) {
+                    let msg = '';
                     if (typeof err.error === 'string') {
                         msg = err.error;
                     } else {