+++ src/util/mergeFunction.js
@@ -1,0 +1,7 @@
+const _ = require('lodash');
+
+module.exports = function(objValue, srcValue) {
+    if (_.isArray(objValue)) {
+        return _.concat(objValue, srcValue);
+    }
+};